+++ NEWS.md
@@ -1,5 +1,7 @@
 # qualcontrol 1.1.0
 * Added `diffvals_summary()`
+* `comparecube` now gets the `any_diffs` column indicating whether any diff column != 0
+* `comparecube` now handles expired and new dimensions
 * Bug fixes
 
 # qualcontrol 1.0.0